+++ src/components/chat-components/SuggestedPrompts.tsx
