+++ src/components/modals/BaseNoteModal.tsx
