+++ src/components/modals/CopilotPlusModal.tsx
