+++ src/components/modals/SimilarNotesModal.tsx
